+++ lib/speed_up_get.dart
@@ -4,5 +4,4 @@ export 'src/extensions.dart';
 export 'src/mixins.dart';
 export 'src/rx_decorator.dart';
 export 'src/app_service.dart';
-export 'src/async_init_service.dart';
 export 'src/functions.dart';