+++ lib/src/async_init_service.dart
@@ -1,5 +1,0 @@
-abstract class IAsyncInitServiceParams {}
-
-abstract class IAsyncInitService<T> {
-  Future initAsync([IAsyncInitServiceParams? params]);
-}